Split Query.Search.cs
To keep file size down, this is now split further by the type of element
being returned by the methods.
